requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
entsoe-py>=0.7.2
pyarrow>=14.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
openpyxl>=3.1.0
gspread>=5.0.0
google-auth>=2.0.0
requests>=2.31.0
